Fix: Update data type in struct from int* to uint8_t*
Updated the data type in the struct from `int*` to `uint8_t*` for better memory management and alignment.

+++ includes/unix_fs_sim/ufs.h
@@ -45,7 +45,7 @@ typedef struct {
  * @brief This struct represents a data stored in the fs.
  */
 typedef struct{
-    int *data;
+    uint8_t *data;
     int position;
     int size_data;
 } data_t;
